requirements.txt
discord
aiohttp
asyncio
mysql-connector-python